Final CR Fixes
Implemented  comments and fixed lint issues
