+++ package.json
@@ -4,6 +4,7 @@
   "homepage": ".",
   "private": true,
   "devDependencies": {
+    "@apollo/client": "^3.3.21",
     "@ethersproject/experimental": "^5.0.1",
     "@fuseio/fuse-swap-default-token-list": "~2.0.0",
     "@fuseio/fuse-swap-sdk": "^3.1.1",
@@ -30,6 +31,7 @@
     "@types/rebass": "^4.0.5",
     "@types/styled-components": "^5.1.0",
     "@types/testing-library__cypress": "^5.0.5",
+    "@types/webfontloader": "^1.6.33",
     "@typescript-eslint/eslint-plugin": "^2.31.0",
     "@typescript-eslint/parser": "^2.31.0",
     "@uniswap/default-token-list": "^1.3.1",
@@ -42,10 +44,12 @@
     "@web3-react/walletconnect-connector": "^6.2.0",
     "@web3-react/walletlink-connector": "^6.0.9",
     "ajv": "^6.12.3",
+    "bignumber.js": "^9.0.1",
     "cids": "^1.0.0",
     "copy-to-clipboard": "^3.2.0",
     "cross-env": "^7.0.2",
     "cypress": "^4.11.0",
+    "dayjs": "^1.10.6",
     "env-cmd": "^10.1.0",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.11.0",
@@ -53,6 +57,7 @@
     "eslint-plugin-react": "^7.19.0",
     "eslint-plugin-react-hooks": "^4.0.0",
     "ethers": "^5.0.7",
+    "graphql": "^15.5.1",
     "i18next": "^15.0.9",
     "i18next-browser-languagedetector": "^3.0.1",
     "i18next-xhr-backend": "^2.0.1",
@@ -65,12 +70,14 @@
     "prettier": "^1.17.0",
     "qs": "^6.9.4",
     "react": "^16.13.1",
+    "react-content-loader": "^6.0.3",
     "react-device-detect": "^1.6.2",
     "react-dom": "^16.13.1",
     "react-feather": "^2.0.8",
     "react-ga": "^2.5.7",
     "react-i18next": "^10.7.0",
     "react-popper": "^2.2.3",
+    "react-promise-tracker": "^2.1.0",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.0.0",
     "react-scripts": "^3.4.1",
@@ -83,8 +90,10 @@
     "serve": "^11.3.0",
     "start-server-and-test": "^1.11.0",
     "styled-components": "^4.2.0",
+    "superagent": "^6.1.0",
     "typescript": "^3.8.3",
-    "use-async-memo": "^1.2.2"
+    "use-async-memo": "^1.2.2",
+    "webfontloader": "^1.6.28"
   },
   "resolutions": {
     "@walletconnect/web3-provider": "1.1.1-alpha.0"
@@ -117,22 +126,5 @@
     ]
   },
   "license": "GPL-3.0-or-later",
-  "dependencies": {
-    "@apollo/react-hooks": "^4.0.0",
-    "@types/webfontloader": "^1.6.33",
-    "apollo-boost": "^0.4.9",
-    "apollo-cache-inmemory": "^1.6.6",
-    "apollo-link-http": "^1.5.17",
-    "bignumber.js": "^9.0.1",
-    "dayjs": "^1.10.6",
-    "event-emitter": "^0.3.5",
-    "graphql": "^15.5.0",
-    "graphql-request": "^3.4.0",
-    "graphql-tag": "^2.11.0",
-    "react-apollo": "^3.1.5",
-    "react-content-loader": "^6.0.3",
-    "react-promise-tracker": "^2.1.0",
-    "superagent": "^6.1.0",
-    "webfontloader": "^1.6.28"
-  }
+  "dependencies": {}
 }
