+++ src/components/Popups/TransactionPopup.tsx
@@ -30,21 +30,23 @@ export default function TransactionPopup({
 
   return (
     <>
-    <RowNoFlex>
-      <div style={{ paddingRight: 16 }}>
-        {success ? <CheckCircle color={theme.green1} size={24} /> : <AlertCircle color={theme.red1} size={24} />}
-      </div>
-      <AutoColumn gap="8px">
-        <TYPE.body fontWeight={500} fontSize={'16px'} color={'#111219'} >{summary ?? 'Hash: ' + hash.slice(0, 8) + '...' + hash.slice(58, 65)}</TYPE.body>
-      </AutoColumn>
-    </RowNoFlex>
-    <Row>
-              {chainId && (
+      <RowNoFlex>
+        <div style={{ paddingRight: 16 }}>
+          {success ? <CheckCircle color={theme.green1} size={24} /> : <AlertCircle color={theme.red1} size={24} />}
+        </div>
+        <AutoColumn gap="8px">
+          <TYPE.body fontWeight={500} fontSize={'16px'} color={'#111219'}>
+            {summary ?? 'Hash: ' + hash.slice(0, 8) + '...' + hash.slice(58, 65)}
+          </TYPE.body>
+        </AutoColumn>
+      </RowNoFlex>
+      <Row>
+        {chainId && (
           <ExternalLink href={getExplorerLink(chainId, hash, 'transaction')}>
             {getExplorerLinkText(chainId)}
           </ExternalLink>
         )}
-    </Row>
-  </>
+      </Row>
+    </>
   )
 }
