+++ src/components/Rewards/apy.tsx
@@ -1,7 +1,8 @@
-import React from 'react'
+import React, { useState } from 'react'
 import styled from 'styled-components'
-import { ShowModal } from '../../hooks/showModal'
-import { Modal } from './modal'
+import { ButtonPrimary } from '../Button'
+import Modal from '../Modal'
+import Questionmark from '../../assets/svg/questionmark.svg'
 
 const Container = styled('div')<{ color: string; txt: string }>`
   display: flex;
@@ -15,12 +16,12 @@ const Container = styled('div')<{ color: string; txt: string }>`
   font-weight: 500;
 `
 
-const Wrap = styled('div')`
+const Wrapper = styled('div')`
   display: flex;
   width: 100%;
   flex-wrap: no-wrap;
 `
-const BorrowApy = styled('div')`
+const Item = styled('div')`
   display: flex;
   width: 100%;
   margin: 3px;
@@ -30,7 +31,7 @@ const BorrowApy = styled('div')`
   }
 `
 
-const LendApy = styled('div')`
+const IconWrapper = styled('div')`
   display: flex;
   width: 100%;
   text-align: center;
@@ -56,7 +57,7 @@ const Icon = styled('div')`
   }
 `
 
-const BorrowLabel = styled('div')`
+const Label = styled('div')`
   display: flex;
   width: 100%;
   text-align: center;
@@ -66,7 +67,7 @@ const BorrowLabel = styled('div')`
     font-weight: 500;
   }
 `
-const LendLabel = styled('div')<{ txt: string }>`
+const Title = styled('div')<{ txt: string }>`
   display: flex;
   width: 100%;
   text-align: center;
@@ -76,33 +77,134 @@ const LendLabel = styled('div')<{ txt: string }>`
   line-height: 18px;
   color: ${({ txt }) => txt};
 `
+export const Backdrop = styled.div`
+  position: fixed;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  background: rgba(0, 0, 0, 0.3);
+  z-index: 500;
+`
+export const StyledModal = styled.div`
+  z-index: 100;
+  padding: 24px;
+  background: #242637;
+  position: relative;
+  margin: auto;
+  border-radius: 12px;
+`
+export const Header = styled.div`
+  border-radius: 8px 8px 0 0;
+  display: flex;
+  justify-content: space-between;
+`
 
-export default function RewardsApy(props: any) {
-  const { isShown, toggle } = ShowModal()
+export const HeaderText = styled.div`
+  color: #fff;
+  align-self: center;
+  color: lightgray;
+`
+
+export const Content = styled.div`
+  padding-bottom: 15px;
+  max-height: 30rem;
+  overflow-x: hidden;
+  overflow-y: auto;
+  > h1 {
+    font-size: 24px;
+    font-weight: 600;
+  }
+  > p {
+    font-size: 14px;
+    font-weight: 400;
+    line-height: 21px;
+  }
+`
+
+const Item2 = styled('div')`
+  display: flex;
+  width: 100%;
+  text-align: center;
+  justify-content: flex-end;
+  position: relative;
+`
+
+const IconW = styled('div')`
+  border-radius: 999px;
+  background-color: #ffffff1a;
+  height: 48px;
+  width: 48px;
+  position: relative;
+  > img {
+    opacity: 0.5;
+    position: absolute;
+    top: 20%;
+    right: 20%;
+    position: absolute;
+  }
+`
+export default function LendingAPY(props: any) {
+  const [isOpen, setOpen] = useState(false)
   const content =
     'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam mi, lorem varius faucibus. Ultricies odio adipiscing integer nunc, quis etiam vehicula lacus. At venenatis elit orci sit diam amet. Vulputate orci id.'
 
   return (
     <Container color={props.color} txt={props.txt}>
-      <Wrap>
-        <BorrowApy>
-          <img src={props.icon} width="18px" height="18px" alt="Currency Icon"></img>{' '}
-        </BorrowApy>
-        <LendApy
+      <Wrapper>
+        <Item>
+          <img src={props.icon} width="18px" height="18px" alt="APY Icon"></img>{' '}
+        </Item>
+        <IconWrapper
           onClick={() => {
-            toggle()
+            setOpen(true)
           }}
         >
           <Icon>
-            <img src={props.apyIcon} width="14px" height="14px" alt="Apy Icon"></img>
+            <img src={props.apyIcon} width="14px" height="14px" alt="APY Icon"></img>
           </Icon>
-        </LendApy>
-      </Wrap>
-      <BorrowLabel>
+        </IconWrapper>
+      </Wrapper>
+      <Label>
         {props.data} &nbsp;<span> {props.label}</span>
-      </BorrowLabel>
-      <LendLabel txt={props.txt}>{props.title}</LendLabel>
-      <Modal headerText={props.title} isShown={isShown} hide={toggle} modalContent={content} />
+      </Label>
+      <Title txt={props.txt}>{props.title}</Title>
+      <Modal
+        maxHeight={90}
+        isOpen={isOpen}
+        onDismiss={() => {
+          setOpen(false)
+        }}
+      >
+        <Wrapper aria-modal aria-labelledby={'APY'} tabIndex={-1} role="dialog">
+          <StyledModal>
+            <Header>
+              <HeaderText>
+                <Item2
+                  onClick={() => {
+                    setOpen(false)
+                  }}
+                >
+                  <IconW>
+                    <img src={Questionmark} width="28px" height="28px" alt="APY Icon"></img>
+                  </IconW>
+                </Item2>
+              </HeaderText>
+            </Header>
+            <Content>
+              <h1>What does &quot; APY &quot; mean?</h1>
+              <p>{content}</p>
+            </Content>
+            <ButtonPrimary
+              onClick={() => {
+                setOpen(false)
+              }}
+            >
+              Done
+            </ButtonPrimary>
+          </StyledModal>
+        </Wrapper>{' '}
+      </Modal>
     </Container>
   )
 }
