+++ src/hooks/useScript.ts
@@ -1,18 +1,16 @@
-import { useEffect } from 'react';
+import { useEffect } from 'react'
 
-export default function useScript(url:string){
-
+export default function useScript(url: string) {
   useEffect(() => {
-    const script = document.createElement('script');
+    const script = document.createElement('script')
 
-    script.src = url;
-    script.async = true;
+    script.src = url
+    script.async = true
 
-    document.body.appendChild(script);
+    document.body.appendChild(script)
 
     return () => {
-      document.body.removeChild(script);
+      document.body.removeChild(script)
     }
-  }, [url]);
-};
-
+  }, [url])
+}
