+++ src/utils/rewards/index.ts
@@ -1,10 +1,10 @@
-import Staking from '../constants/abis/stakeMethods.json'
+import Staking from '../../constants/abis/stakeMethods.json'
 import ethers from 'ethers'
 import { formatEther, parseUnits } from 'ethers/lib/utils'
-import { getProviderOrSigner } from '.'
-import BasicTokenABI from '../constants/abis/tokenABI.json'
+import { getProviderOrSigner } from '..'
+import BasicTokenABI from '../../constants/abis/tokenABI.json'
 
-var url = 'https://rpc.fuse.io'
+const url = 'https://rpc.fuse.io'
 
 export async function getRewardsData(contractAddress: string, LP: any, account: any) {
   const provider = new ethers.providers.JsonRpcProvider(url)
